Implement NoDefaultIncludedAttribute
