+++ Saule/Http/AllowsQueryAttribute.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Net.Http;
 using System.Web.Http.Controllers;
 using System.Web.Http.Filters;
-using Saule.Queries;
 using Saule.Queries.Filtering;
 using Saule.Queries.Including;
 using Saule.Queries.Sorting;
@@ -29,32 +28,13 @@ namespace Saule.Http
             var filtering = new FilteringContext(queryParams);
             var including = new IncludingContext(queryParams);
 
-            var queryContext = GetQueryContext(actionContext);
+            var queryContext = QueryContextUtils.GetQueryContext(actionContext);
 
             queryContext.Sorting = sorting;
             queryContext.Filtering = filtering;
             queryContext.Including = including;
 
             base.OnActionExecuting(actionContext);
-        }
-
-        private static QueryContext GetQueryContext(HttpActionContext actionContext)
-        {
-            var hasQuery = actionContext.Request.Properties.ContainsKey(Constants.PropertyNames.QueryContext);
-            QueryContext query;
-
-            if (hasQuery)
-            {
-                query = actionContext.Request.Properties[Constants.PropertyNames.QueryContext]
-                    as QueryContext;
-            }
-            else
-            {
-                query = new QueryContext();
-                actionContext.Request.Properties.Add(Constants.PropertyNames.QueryContext, query);
-            }
-
-            return query;
         }
     }
 }
