+++ Saule/Queries/QueryContext.cs
@@ -14,5 +14,7 @@ namespace Saule.Queries
         public FilteringContext Filtering { get; set; }
 
         public IncludingContext Including { get; set; }
+
+        public bool IncludedDefault { get; set; }
     }
 }
